+++ test/cookbooks/test/metadata.rb
@@ -3,3 +3,4 @@ version '0.1.0'
 
 depends 'git'
 depends 'nodejs'
+depends 'yum-centos'